Added a system clock to handle timing and cycles.
Updated all cpu instructions to work alongside the clock, taking into account how many cycles an instruction takes.

Updated tests so tests are available for all cpu instructions.

Added the use of threading so the clock works on 1 thread and the cpu works on another thread, to parallelize the 2 components.

Added the `__bool__` method to DType.